+++ aruco_detect/package.xml
@@ -25,5 +25,9 @@
   <depend>dynamic_reconfigure</depend>
   <depend>python-cairosvg</depend>
   <depend>python-joblib</depend>
+  <depend>nodelet</depend>
 
+  <export>
+    <nodelet plugin="${prefix}/nodelet_plugins.xml"/>
+  </export>
 </package>
